refactor: decouple lesson creation & unique-ication from scraper
