+++ common/src/model/Lesson.ts
@@ -115,6 +115,50 @@ export const createLessonWithParticipantsId = (lesson: NonUniqueLesson): string 
 	return id;
 };
 
+export const createLessonWithParticipants = (data: LessonInitData = getDefaultNonUniqueLesson()): NonUniqueLesson => {
+	const { participants } = data;
+
+	const teachers = p2teachers(participants);
+	const rooms = p2rooms(participants);
+
+	/** @deprecated - only for backwards compatibility */
+	const teacher = teachers.join(", ").trim();
+
+	/** @deprecated - only for backwards compatibility */
+	const room = rooms.join(", ").trim();
+
+	const almostLesson: Omit<NonUniqueLesson, "id"> = {
+		...data,
+
+		teacher,
+		room,
+	};
+
+	const id: string = createLessonWithParticipantsId(almostLesson as NonUniqueLesson);
+
+	return {
+		...almostLesson,
+		id,
+	};
+};
+
+export const createLesson = (data: NonUniqueLesson = getDefaultNonUniqueLesson()): Lesson => {
+	const { participants, ...rest } = data;
+
+	const teachers = p2teachers(participants);
+	const rooms = p2rooms(participants);
+	const classes = p2classes(participants);
+	const students = p2students(participants);
+
+	return {
+		...rest,
+		teachers,
+		rooms,
+		classes,
+		students,
+	};
+};
+
 /**
  * sorts the lessons array,
  * first by the day index,
@@ -171,3 +215,13 @@ const mergeDuplicateLessonsStrat: MergeStrategy<NonUniqueLesson> = (left, right)
  *
  */
 export const mergeDuplicateLessons = mergeBy("id", mergeDuplicateLessonsStrat);
+
+export const makeLessonsUnique = (nonUniqueLessons: NonUniqueLesson[] = []): Lesson[] => {
+	const merged = mergeDuplicateLessons(nonUniqueLessons);
+
+	const sorted = sortLessons(merged);
+
+	const finished = sorted.map(createLesson);
+
+	return finished;
+};
